A new version that supports the new esp32-c5 (prototype--it's still in the mail)
still builds on esp32. HOPEFULLY builds on the c5, but not yet tested (it's in the mail)
I realized there are some issues introduced with proportions of device types and beacons vs probes. 
built out the more convincing devices and reduced the junk a little bit but still try to keep the radio running at 100%
Added RAM protections. so far I haven't seen them really required, but in a larger crowd I think it'll be required because we'll be learning way more ssids.