Update ghostwalk5ghz.cpp - builds on c5
I had written the c5 code theoretically. now I have the device. had to make some changes so it would build. 
The c5 variant does not support the tft, but if it's a standard esp32 it does. there's no mass produced tft c5 i'm aware of. tft info is redirected to serial. tft is mostly for development anyway.